refactor: add buffer interface
